Fine tuning for display of userlist and tunelist
Basically, on big screens we want everything centered whereas on
small screens we want everything flush to the LHS so that we don't
waste real estate. This applies to both lists.

Because thumbnails occupy so much space, clip both the tune title
and the thumbnail on small screens (but not on big ones).  Where
they are not selected, allocate more space to the tune title.

Don't delay for quite so long when rendering thumbnails in order
for the refresh of the screen to catch up.